% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_interpret.R
\name{safer_eval}
\alias{safer_eval}
\title{Evaluate an expression if functions are whitelisted}
\usage{
safer_eval(expr, whitelist)
}
\arguments{
\item{expr}{an expression to evaluate}

\item{whitelist}{list or character vector of allowed functions}
}
\value{
The return value from evaluating the expression
}
\description{
Evaluate an expression if functions are whitelisted
}
\references{
https://community.rstudio.com/t/secure-way-to-accept-user-input-for-shiny-apps/13554
}
