% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_interpret.R
\name{interpret}
\alias{interpret}
\title{Interpret input string and evaluate if not-too-spicy}
\usage{
interpret(
  expr_str,
  max_length = 32,
  whitelist = c("/", "*", "^", "+", "sqrt", "(", ")")
)
}
\arguments{
\item{expr_str}{a character string containing an expression to evaluate}

\item{max_length}{hard limit on expression string length}

\item{whitelist}{list or character vector functions allowed in \code{expr}}
}
\description{
A utility to interpret calculator input
}
